Change isTest handling so that isTest=false and no p1-p5 parameters will result in showing a tutor that has no test controls and not show problems marked testable.   This meant creating having to temp user types:  externalTempCaller and externalTempTester.   The caller gets no test controls and only sees ready content.   The tester gets controls and sees testable content in addition to ready content.   Neither updates stats.   This allows testing the system and seeing it as a real user would see it.     I've left the default to be isTest=true if this param is not provided.
Discovered a bug that is handled by reporting an exception.  If a non test user tries to teachTopic and that topic has no ready problems (all are testable) then we can't show anything.   The best we can do is return an exception reporting that you need to call the system with isTest=true if there arent any ready problems in a topic