This is a very long comment due to the fact that my email is currently down, and this is now my only way to communicate this information.
Dynamic pedagogies in this commit not causing any bugs that I can find.  However, it is also not fully finished in this version.  I thought I should commit the last fully bug-free code now, so other people could see what is there, as the changes are turning out to be rather extensive.

Still to-do: redo logging with better format (removed entirelt at the moment)
Make pop-up notifications for testing
Figure out why the learning companion updates behind the scenes but not visually (discovered this today, since everything looked good server side and in the logs).  I either need to move the change earlier or force an update.